third_party/blink/web_tests/crypto/subtle/aes-kw/key-manipulation-expected.txt

Test generating, importing and exporting keys for AES-KW. Test that they can't be used with another algorithm.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Generating a key...
SUCCESS
PASS key.toString() is '[object CryptoKey]'
PASS key.type is 'secret'
PASS key.algorithm.name is 'AES-KW'
PASS key.algorithm.length is 256

Testing that the key can't be used with AES-CBC...
SUCCESS (rejected): InvalidAccessError: key.algorithm does not match that of operation

Exporting the key to raw...
SUCCESS
PASS exportedKey.toString() is '[object ArrayBuffer]'

Importing it back...
SUCCESS
PASS importedKey.toString() is '[object CryptoKey]'
PASS importedKey.type is 'secret'
PASS importedKey.algorithm.name is 'AES-KW'
PASS importedKey.algorithm.length is 256
PASS successfullyParsed is true

TEST COMPLETE
